pyproject.toml
[tool.poetry]
name = "zerocode"
version = "0.1.0"
description = ""
authors = ["Vladislav-76 <sillmaril2@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyTelegramBotAPI = "^4.18.1"
requests = "^2.32.2"
openai = "^1.30.4"
python-dotenv = "^1.0.1"
gtts = "^2.5.1"
speechrecognition = "^3.10.4"
pydub = "^0.25.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
